+++ packages/libraryB/CHANGELOG.md
@@ -1,35 +1,22 @@
-## 0.0.2-aplha
+## 0.0.3-aplha
 
-Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
+Associated Task: [#265233](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=26523324-hanshijie)
 
 ### ✨Features
 
-Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
+Associated Task: [#265233](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=26523324-hanshijie)
 
-- [9288c3e](http://github.com/commit/9288c3efd5e7efcf7d4a7d31232a34223de74464) Thanks [ncqwer](https://github.com/ncqwer) ! -  wowo
-- [7064d31](http://github.com/commit/7064d313a8d569888e8986d135adc6670cafa859) Thanks [ncqwer](https://github.com/ncqwer) ! -  hahaha
-- [3c976f2](http://github.com/commit/3c976f2e4482879b7e28a4ff704b46cd708d3f45) Thanks [ncqwer](https://github.com/ncqwer) ! -  just
-- [80136ca](http://github.com/commit/80136cac5da7fb3308c7420be7748003858fd557) Thanks [ncqwer](https://github.com/ncqwer) ! -  不知道干嘛
-- [4283536](http://github.com/commit/42835362e3f4bf33c365865ca28a13c0389d5af0) Thanks [ncqwer](https://github.com/ncqwer) ! -  hello world
-- [0659d3d](http://github.com/commit/0659d3d165225949a5c8c8af2bce83cc1f7afb77) Thanks [ncqwer](https://github.com/ncqwer) ! -  wowo
-- [e794ed1](http://github.com/commit/e794ed1d88512b65907efa64133b801b84d6226d) Thanks [ncqwer](https://github.com/ncqwer) ! -  hahaha
-- [09c1233](http://github.com/commit/09c12335dd0461573197a3d855be6eb2bd37fe84) Thanks [ncqwer](https://github.com/ncqwer) ! -  just
-- [2795505](http://github.com/commit/27955050f1b2f236bed687084e6b3c13a7460b1c) Thanks [ncqwer](https://github.com/ncqwer) ! -  不知道干嘛
-- [b439188](http://github.com/commit/b439188f41e29d5d39bbf9764af4c1a1c3d0753f) Thanks [ncqwer](https://github.com/ncqwer) ! -  hello world
-- [b512329](http://github.com/commit/b512329e663733d86d9230078dd544cf2b642dc1) Thanks [ncqwer](https://github.com/ncqwer) ! -  just
-- [c0e8755](http://github.com/commit/c0e87559c9184513366db0fe26ea45ff81657e6e) Thanks [ncqwer](https://github.com/ncqwer) ! -  不知道干嘛
-- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! -  hello world
+- [0719f3b](https://github.com/ncqwer/github_action_playground/commit/0719f3b2039b7dc80ee3a5593f5c02154b8701aa) Thanks [ncqwer](https://github.com/ncqwer) ! - 删去多余依赖
 
+## 0.0.2-aplha
 
-### 🚨BREAKING CHANGES
+Associated Task: [#265233](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=26523324-hanshijie)
 
-Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
+### ✨Features
 
-- [4283536](http://github.com/commit/42835362e3f4bf33c365865ca28a13c0389d5af0) Thanks [ncqwer](https://github.com/ncqwer) ! - ceshi
-- [b439188](http://github.com/commit/b439188f41e29d5d39bbf9764af4c1a1c3d0753f) Thanks [ncqwer](https://github.com/ncqwer) ! - ceshi
-- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! - ceshi
+Associated Task: [#265233](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=26523324-hanshijie)
 
-
+- [0719f3b](https://github.com/ncqwer/github_action_playground/commit/0719f3b2039b7dc80ee3a5593f5c02154b8701aa) Thanks [ncqwer](https://github.com/ncqwer) ! - 删去多余依赖
 
 ## 0.0.1-aplha
 
@@ -39,19 +26,16 @@ Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.co
 
 Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
 
-- [1a78bd4](http://github.com/commit/1a78bd4d64c7e1a4b7dfe3c5bcca3c790a41fdb5) Thanks [ncqwer](https://github.com/ncqwer) ! -  hahaha
-- [b512329](http://github.com/commit/b512329e663733d86d9230078dd544cf2b642dc1) Thanks [ncqwer](https://github.com/ncqwer) ! -  just
-- [c0e8755](http://github.com/commit/c0e87559c9184513366db0fe26ea45ff81657e6e) Thanks [ncqwer](https://github.com/ncqwer) ! -  不知道干嘛
-- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! -  hello world
-
+- [1a78bd4](http://github.com/commit/1a78bd4d64c7e1a4b7dfe3c5bcca3c790a41fdb5) Thanks [ncqwer](https://github.com/ncqwer) ! - hahaha
+- [b512329](http://github.com/commit/b512329e663733d86d9230078dd544cf2b642dc1) Thanks [ncqwer](https://github.com/ncqwer) ! - just
+- [c0e8755](http://github.com/commit/c0e87559c9184513366db0fe26ea45ff81657e6e) Thanks [ncqwer](https://github.com/ncqwer) ! - 不知道干嘛
+- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! - hello world
 
 ### 🚨BREAKING CHANGES
 
 Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
 
 - [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! - ceshi
-
-
 
 ## 0.0.2
 
@@ -61,17 +45,13 @@ Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.co
 
 Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
 
-- [1a78bd4](http://github.com/commit/1a78bd4d64c7e1a4b7dfe3c5bcca3c790a41fdb5) Thanks [ncqwer](https://github.com/ncqwer) ! -  hahaha
-- [b512329](http://github.com/commit/b512329e663733d86d9230078dd544cf2b642dc1) Thanks [ncqwer](https://github.com/ncqwer) ! -  just
-- [c0e8755](http://github.com/commit/c0e87559c9184513366db0fe26ea45ff81657e6e) Thanks [ncqwer](https://github.com/ncqwer) ! -  不知道干嘛
-- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! -  hello world
-
+- [1a78bd4](http://github.com/commit/1a78bd4d64c7e1a4b7dfe3c5bcca3c790a41fdb5) Thanks [ncqwer](https://github.com/ncqwer) ! - hahaha
+- [b512329](http://github.com/commit/b512329e663733d86d9230078dd544cf2b642dc1) Thanks [ncqwer](https://github.com/ncqwer) ! - just
+- [c0e8755](http://github.com/commit/c0e87559c9184513366db0fe26ea45ff81657e6e) Thanks [ncqwer](https://github.com/ncqwer) ! - 不知道干嘛
+- [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! - hello world
 
 ### 🚨BREAKING CHANGES
 
 Associated Task: [#265652](https://projectmanage.netease-official.lcap.163yun.com/dashboard/TaskDetail?id=2656521174974208)
 
 - [a7dda74](http://github.com/commit/a7dda745af1dc9e1c1e62e18f1de619a782713a0) Thanks [ncqwer](https://github.com/ncqwer) ! - ceshi
-
-
-
