+++ packages/libraryB/package.json
@@ -1,4 +1,4 @@
 {
   "name": "mono-example",
-  "version": "0.0.2-aplha"
+  "version": "0.0.3-aplha"
 }
